requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.38.0
timm>=0.9.0
Pillow>=9.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.3.0
open-clip-torch>=2.20.0

# Comment agent
anthropic>=0.25.0

# Demo app
gradio>=4.0.0
requests>=2.31.0